pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "django==5.2",
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.3.2",
    "python-decouple>=3.8",
    "celery>=5.4.0",
    "channels>=4.0.0",
    "channels-redis>=4.2.0",
    "daphne>=4.1.0",
    "djangorestframework-simplejwt>=5.4.0",
    "model-bakery>=1.20.0",
    "pytest-django>=4.9.0",
    "pytest-asyncio>=0.24.0",
    "isort>=7.0.0",
    "flake8>=7.3.0",
    "pre-commit>=4.5.1",
    "drf-spectacular>=0.29.0",
]


[tool.black]
line-length = 119
target-version = ['py311']
exclude = '''/( \.git | \.venv | \.vscode | \.pytest_cache | \.env | env | venv | pyproject\.toml)/'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 119
skip = [".env", "venv", "env", ".venv"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
python_files = ["test_*.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: Testes rápidos de unidade (models, services)",
    "integration: Testes de integração (websockets, api, tasks)",
]
